Move RecoveryPhrase under manageaccount
