+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/backupkey/BackupKeyViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.entities.Account
 import io.horizontalsystems.bankwallet.entities.AccountType
-import io.horizontalsystems.bankwallet.modules.recoveryphrase.RecoveryPhraseModule
+import io.horizontalsystems.bankwallet.modules.manageaccount.recoveryphrase.RecoveryPhraseModule
 
 class BackupKeyViewModel(val account: Account) : ViewModel() {
