+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/manageaccount/ManageAccountFragment.kt
@@ -33,7 +33,7 @@ import io.horizontalsystems.bankwallet.modules.balance.ui.NoteWarning
 import io.horizontalsystems.bankwallet.modules.manageaccount.ManageAccountModule.ACCOUNT_ID_KEY
 import io.horizontalsystems.bankwallet.modules.manageaccount.ManageAccountModule.KeyAction
 import io.horizontalsystems.bankwallet.modules.manageaccount.publickeys.PublicKeysModule
-import io.horizontalsystems.bankwallet.modules.recoveryphrase.RecoveryPhraseModule
+import io.horizontalsystems.bankwallet.modules.manageaccount.recoveryphrase.RecoveryPhraseModule
 import io.horizontalsystems.bankwallet.modules.unlinkaccount.UnlinkAccountDialog
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.TranslatableString
